test: consolidate wire-compat tests for actions, services, and topics into a single module
