+++ crates/generator-internal/tests/python.rs
@@ -3,11 +3,11 @@ mod helpers;
 
 #[path = "python/actions_communication.rs"]
 mod actions_communication;
-#[path = "python/actions_wire_compat.rs"]
-mod actions_wire_compat;
 #[path = "python/generate_lib.rs"]
 mod generate_lib;
 #[path = "python/services_communication.rs"]
 mod services_communication;
 #[path = "python/topics_communication.rs"]
 mod topics_communication;
+#[path = "python/wire_compat.rs"]
+mod wire_compat;
